+++ tests/test_cli_environment.py
@@ -31,9 +31,12 @@ class ArgumentParsingTest(OutputCaptureTest):
             "arg_value": "" // conf_utils.Cmdline(arg="--arg-value"),
             "list": ["existing_item"] // conf_utils.Cmdline(append="--append"),
         })
+        self._parser = cli_utils.SlashArgumentParser()
+        cli_utils.configure_arg_parser_by_config(self._parser, self.config)
 
     def _cli(self, argv):
-        return cli_utils.get_cli_environment_context(argv=argv, config=self.config)
+        parsed_args, _ = self._parser.parse_known_args(argv)
+        return cli_utils.get_modified_configuration_from_args_context(self._parser, parsed_args, config=self.config)
 
     def test_config_arg_flag(self):
         with self._cli(["--arg-value=x"]):
@@ -83,31 +86,42 @@ class PluginCommandLineArgumentsTest(OutputCaptureTest):
         self.plugin = SampleCommandLinePlugin()
         plugins.manager.install(self.plugin)
         self.addCleanup(plugins.manager.uninstall, self.plugin)
+        self._parser = cli_utils.SlashArgumentParser()
 
     def test_arguments_are_not_parsed_if_not_activated(self):
         args = ["--start-session-option", "2"]
-        with self.assertRaises(SystemExit):
-            with cli_utils.get_cli_environment_context(argv=args):
-                pass
+        with self.assertRaises(SystemExit) as caught:
+            self._parser.parse_args(args)
+        assert caught.exception.code != 0
 
     def test_activation(self):
-        with cli_utils.get_cli_environment_context(argv=["--with-sample-plugin"]):
-            self.assertIn(self.plugin.get_name(), plugins.manager.get_active_plugins(), "plugin was not activated")
+        cli_utils.add_pending_plugins_from_commandline(["--with-sample-plugin"])
+        self.assertIn(self.plugin.get_name(), plugins.manager.get_future_active_plugins(), "plugin was not activated")
 
     def test_deactivation(self):
+        argv = ["--without-sample-plugin"]
         plugins.manager.activate(self.plugin)
-        with cli_utils.get_cli_environment_context(argv=["--without-sample-plugin"]):
-            self.assertNotIn(self.plugin.get_name(), plugins.manager.get_active_plugins())
+        argv = cli_utils.add_pending_plugins_from_commandline(argv)
+        self.assertIn(self.plugin.get_name(), plugins.manager.get_active_plugins())
+        assert argv == []
+        plugins.manager.activate_pending_plugins()
+        self.assertNotIn(self.plugin.get_name(), plugins.manager.get_future_active_plugins())
         self.assertNotIn(self.plugin.get_name(), plugins.manager.get_active_plugins())
 
+
     def test_argument_passing(self):
-        with cli_utils.get_cli_environment_context(argv=["--with-sample-plugin", "--plugin-option", "value"]):
-            self.assertEquals(self.plugin.cmdline_param, "value")
+        argv = ["--with-sample-plugin", "--plugin-option", "value"]
+        cli_utils.configure_arg_parser_by_plugins(self._parser)
+        argv = cli_utils.add_pending_plugins_from_commandline(argv)
+        parsed_args = self._parser.parse_args(argv)
+        plugins.manager.activate_pending_plugins()
+        cli_utils.configure_plugins_from_args(parsed_args)
+        self.assertEqual(self.plugin.cmdline_param, "value")
 
     def test_help_shows_available_plugins(self):
+        cli_utils.configure_arg_parser_by_plugins(self._parser)
         with self.assertRaises(SystemExit):
-            with cli_utils.get_cli_environment_context(argv=["-h"]):
-                pass
+            self._parser.parse_args(['-h'])
         output = self.stdout.getvalue()
         self.assertIn("--with-sample-plugin", output)