+++ doc/changelog.rst
@@ -2,6 +2,7 @@ Changelog
 =========
 
 * :feature:`519` Add ``--no-output`` flag for ``slash list``
+* :feature:`497` Major overhaul of CLI mechanics -- improve help message and usage, as well as cleaner error exits during the session configuration phase
 * :feature:`467` Yield fixtures are now automatically detected by Slash -- using ``yield_fixture`` explicitly is no longer required.
 * :feature:`507` Test id can now be obtained via ``slash.context.test.id``
 * :bug:`510 major` Explicitly fail fixtures which name is valid for tests (currently: ``test_`` prefix)